Larger internal refactoring
- This refactoring should be largely backwards compatible as long as no
  internal JGiven classes have been used
- Created a new interface StageCreator to make it easier to change the
  stage creation strategy. This simplifies the Spring and the Android
  integrations
- Changed the way the StepInterceptor is used
- Created new classes JGivenMethodRule and JGivenClassRule to replace
  the old rule classes. The old classes are deprecated now
- Simplified the implementation of the Spring and Android integrations
