+++ jgiven-android/src/main/java/com.tngtech.jgiven.integration.android/AndroidStageCreator.java
@@ -1,0 +1,22 @@
+package com.tngtech.jgiven.integration.android;
+
+import android.support.test.InstrumentationRegistry;
+import android.support.v4.content.ContextCompat;
+
+import com.tngtech.jgiven.impl.ByteBuddyStageCreator;
+
+import net.bytebuddy.android.AndroidClassLoadingStrategy;
+import net.bytebuddy.dynamic.loading.ClassLoadingStrategy;
+
+public class AndroidStageCreator extends ByteBuddyStageCreator {
+
+    @Override
+    protected ClassLoader getClassLoader(Class<?> stageClass) {
+        return InstrumentationRegistry.getTargetContext().getClassLoader();
+    }
+
+    @Override
+    protected ClassLoadingStrategy getClassLoadingStrategy(Class<?> stageClass) {
+        return new AndroidClassLoadingStrategy(ContextCompat.getCodeCacheDir(InstrumentationRegistry.getTargetContext()));
+    }
+}
