+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/StageCreator.java
@@ -1,0 +1,11 @@
+package com.tngtech.jgiven.impl;
+
+import com.tngtech.jgiven.impl.intercept.StepInterceptor;
+
+/**
+ * Creates instances of stage classes
+ */
+public interface StageCreator {
+
+    <T> T createStage(Class<T> stageClass,  StepInterceptor stepInterceptor);
+}
