+++ jgiven-core/src/main/java/com/tngtech/jgiven/impl/intercept/ByteBuddyMethodInterceptor.java
@@ -1,26 +1,28 @@
 package com.tngtech.jgiven.impl.intercept;
 
+import com.tngtech.jgiven.impl.intercept.StepInterceptor.Invoker;
+
 import java.lang.reflect.Method;
 
 import java.util.concurrent.Callable;
 import net.bytebuddy.implementation.bind.annotation.AllArguments;
 import net.bytebuddy.implementation.bind.annotation.BindingPriority;
 import net.bytebuddy.implementation.bind.annotation.DefaultCall;
-import net.bytebuddy.implementation.bind.annotation.Morph;
 import net.bytebuddy.implementation.bind.annotation.Origin;
 import net.bytebuddy.implementation.bind.annotation.RuntimeType;
 import net.bytebuddy.implementation.bind.annotation.SuperCall;
 import net.bytebuddy.implementation.bind.annotation.This;
 
 /**
- * StepMethodInterceptor that uses ByteBuddy Method interceptor with annotations for intercepting JGiven methods
+ * StepInterceptorImpl that uses ByteBuddy Method interceptor with annotations for intercepting JGiven methods
  *
  */
-public class StandaloneStepMethodInterceptor extends StepMethodInterceptor {
+public class ByteBuddyMethodInterceptor {
 
-    public StandaloneStepMethodInterceptor(
-        StepMethodHandler scenarioMethodHandler, StageTransitionHandler stageStateUpdater ) {
-        super( scenarioMethodHandler, stageStateUpdater );
+    private final StepInterceptor interceptor;
+
+    public ByteBuddyMethodInterceptor(StepInterceptor interceptor ) {
+        this.interceptor = interceptor;
     }
 
     @RuntimeType
@@ -29,14 +31,13 @@ public class StandaloneStepMethodInterceptor extends StepMethodInterceptor {
         Method method, @AllArguments final Object[] parameters)
         throws Throwable {
         Invoker invoker = new Invoker() {
-
             @Override
             public Object proceed() throws Throwable {
                 return zuper.call();
             }
 
         };
-        return doIntercept( receiver  , method, parameters, invoker );
+        return interceptor.intercept( receiver  , method, parameters, invoker );
     }
 
     @RuntimeType
@@ -52,7 +53,7 @@ public class StandaloneStepMethodInterceptor extends StepMethodInterceptor {
             }
 
         };
-        return doIntercept( receiver  , method, parameters, invoker );
+        return interceptor.intercept( receiver  , method, parameters, invoker );
     }
     @RuntimeType
     public Object intercept( @This final Object receiver,@Origin final Method method, @AllArguments final Object[] parameters) throws Throwable {
@@ -65,7 +66,7 @@ public class StandaloneStepMethodInterceptor extends StepMethodInterceptor {
             }
 
         };
-        return doIntercept( receiver  , method, parameters, invoker );
+        return interceptor.intercept( receiver  , method, parameters, invoker );
     }
