+++ jgiven-junit/src/main/java/com/tngtech/jgiven/junit/de/SzenarioTest.java
@@ -4,16 +4,16 @@ import com.tngtech.jgiven.impl.Scenario;
 import org.junit.ClassRule;
 import org.junit.Rule;
 
-import com.tngtech.jgiven.junit.ScenarioExecutionRule;
-import com.tngtech.jgiven.junit.ScenarioReportRule;
+import com.tngtech.jgiven.junit.JGivenMethodRule;
+import com.tngtech.jgiven.junit.JGivenClassRule;
 import com.tngtech.jgiven.lang.de.SzenarioTestBasis;
 
 public class SzenarioTest<GEGEBEN, WENN, DANN> extends SzenarioTestBasis<GEGEBEN, WENN, DANN> {
     @ClassRule
-    public static final ScenarioReportRule writerRule = new ScenarioReportRule();
+    public static final JGivenClassRule writerRule = new JGivenClassRule();
 
     @Rule
-    public final ScenarioExecutionRule scenarioRule = new ScenarioExecutionRule( createScenario() );
+    public final JGivenMethodRule scenarioRule = new JGivenMethodRule( createScenario() );
 
     @Override
     public Scenario<GEGEBEN, WENN, DANN> getScenario() {
