+++ jgiven-junit/src/test/java/com/tngtech/jgiven/junit/injection/ComposedStepInjectionTest.java
@@ -2,15 +2,18 @@ package com.tngtech.jgiven.junit.injection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.junit.Rule;
 import org.junit.Test;
 
 import com.tngtech.jgiven.Stage;
 import com.tngtech.jgiven.annotation.ExpectedScenarioState;
 import com.tngtech.jgiven.annotation.ScenarioStage;
 import com.tngtech.jgiven.annotation.ScenarioState;
-import com.tngtech.jgiven.junit.ScenarioTest;
+import com.tngtech.jgiven.junit.JGivenMethodRule;
 
-public class ComposedStepInjectionTest extends ScenarioTest {
+public class ComposedStepInjectionTest {
+    @Rule
+    public final JGivenMethodRule scenarioRule = new JGivenMethodRule();
 
     @ScenarioStage
     CustomerSteps customerSteps;
