+++ jgiven-junit/src/test/java/com/tngtech/jgiven/junit/injection/MultilevelComposedStepInjectionTest.java
@@ -2,14 +2,18 @@ package com.tngtech.jgiven.junit.injection;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import org.junit.Rule;
 import org.junit.Test;
 
 import com.tngtech.jgiven.Stage;
 import com.tngtech.jgiven.annotation.ExpectedScenarioState;
 import com.tngtech.jgiven.annotation.ScenarioStage;
-import com.tngtech.jgiven.junit.ScenarioTest;
+import com.tngtech.jgiven.junit.JGivenMethodRule;
 
-public class MultilevelComposedStepInjectionTest extends ScenarioTest {
+public class MultilevelComposedStepInjectionTest {
+
+    @Rule
+    public final JGivenMethodRule scenarioRule = new JGivenMethodRule();
 
     @ScenarioStage
     CustomerSteps customerSteps;
