+++ jgiven-spring/src/main/java/com/tngtech/jgiven/integration/spring/SimpleSpringScenarioTest.java
@@ -1,14 +1,9 @@
 package com.tngtech.jgiven.integration.spring;
 
 import com.tngtech.jgiven.junit.SimpleScenarioTest;
-import org.junit.ClassRule;
-import org.junit.Rule;
+
 import org.springframework.beans.factory.BeanFactory;
 import org.springframework.beans.factory.BeanFactoryAware;
-
-import com.tngtech.jgiven.base.SimpleScenarioTestBase;
-import com.tngtech.jgiven.junit.ScenarioExecutionRule;
-import com.tngtech.jgiven.junit.ScenarioReportRule;
 
 /**
  * A variant of {@link com.tngtech.jgiven.integration.spring.SpringScenarioTest} works with a single
@@ -19,7 +14,7 @@ import com.tngtech.jgiven.junit.ScenarioReportRule;
 public class SimpleSpringScenarioTest<STAGE> extends SimpleScenarioTest<STAGE> implements BeanFactoryAware {
 
     public void setBeanFactory( BeanFactory beanFactory ) {
-        getScenario().setExecutor( beanFactory.getBean( SpringScenarioExecutor.class ) );
+        getScenario().setStageCreator( beanFactory.getBean( SpringStageCreator.class ) );
     }
 
 }
