+++ jgiven-spring/src/main/java/com/tngtech/jgiven/integration/spring/SpringStageCreator.java
@@ -9,8 +9,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.ApplicationContext;
 import org.springframework.util.ClassUtils;
 
-import com.tngtech.jgiven.impl.ScenarioExecutor;
-import com.tngtech.jgiven.impl.StandaloneScenarioExecutor;
+import com.tngtech.jgiven.impl.ByteBuddyStageCreator;
+import com.tngtech.jgiven.impl.intercept.StepInterceptor;
 
 /**
  * Main class of JGiven for executing scenarios with spring support.
@@ -20,23 +20,23 @@ import com.tngtech.jgiven.impl.StandaloneScenarioExecutor;
  * <pre>
  *	{@literal @}Bean
  *	{@literal @}Scope("prototype")
- *	public SpringScenarioExecutor springScenarioExecutor() {
- *	    return new SpringScenarioExecutor();
+ *	public SpringStageCreator springScenarioExecutor() {
+ *	    return new SpringStageCreator();
  *	}
  * </pre>
  * <p>
- * <strong>The SpringScenarioExecutor is stateful, and thus should use "prototype" scope</strong>
+ * <strong>The SpringStageCreator is stateful, and thus should use "prototype" scope</strong>
  * @since 0.8.0
  */
-public class SpringScenarioExecutor extends StandaloneScenarioExecutor implements ScenarioExecutor {
+public class SpringStageCreator extends ByteBuddyStageCreator {
 
-    private static final Logger log = LoggerFactory.getLogger( SpringScenarioExecutor.class );
+    private static final Logger log = LoggerFactory.getLogger( SpringStageCreator.class );
 
     @Autowired
     private ApplicationContext applicationContext;
 
     @Override
-    public <T> T createStageClass( Class<T> stepsClass ) {
+    public <T> T createStage(Class<T> stepsClass, StepInterceptor stepInterceptor ) {
         try {
             T bean = applicationContext.getBean( stepsClass );
             Advised advised = (Advised) bean;
@@ -44,25 +44,17 @@ public class SpringScenarioExecutor extends StandaloneScenarioExecutor implement
             for( Advisor advisor : advisors ) {
                 if( advisor.getAdvice() instanceof SpringStepMethodInterceptor ) {
                     SpringStepMethodInterceptor interceptor = (SpringStepMethodInterceptor) advisor.getAdvice();
-                    interceptor.setScenarioMethodHandler( methodHandler );
-                    interceptor.setStageTransitionHandler( stageTransitionHandler );
-                    interceptor.enableMethodHandling( true );
+                    interceptor.setStepInterceptor(stepInterceptor);
                 }
             }
             return bean;
         } catch( NoSuchBeanDefinitionException nbe ) {
-            return super.createStageClass( stepsClass );
+            return super.createStage( stepsClass, stepInterceptor );
         } catch( ClassCastException cce ) {
-            log.error( "class " + ClassUtils.getShortName( stepsClass )
+            log.warn( "Class " + ClassUtils.getShortName( stepsClass )
                     + " is not advised with SpringStepMethodInterceptor. Falling back to cglib based proxy, strange things may happen." );
-            return super.createStageClass( stepsClass );
+            return super.createStage( stepsClass, stepInterceptor );
         }
-    }
-
-    @Override
-    public StageState getStageState( Object stage ) {
-        StageState stageState = stages.get( stage.getClass() );
-        return stageState != null ? stageState : super.getStageState( stage );
     }
 
 }
